+++ Heroku_app/names.py
@@ -1,0 +1,40 @@
+STOCKS = {
+    'Alibaba': 'BABA',
+    'Alphabet': 'GOOGL',
+    'Amazon': 'AMZN',
+    'Apple': 'AAPL',
+    'Facebook': 'FB',
+    'Microsoft': 'MSFT',
+    'Netflix': 'NFLX',
+    'Oracle': 'ORCL',
+    'Tesla': 'TSLA',
+    'Twitter': 'TWTR'
+}
+STOCKS = {
+    'Alibaba': 'BABA',
+    'Alphabet': 'GOOGL',
+    'Amazon': 'AMZN',
+    'Apple': 'AAPL',
+    'Facebook': 'FB',
+    'Microsoft': 'MSFT',
+    'Netflix': 'NFLX',
+    'Oracle': 'ORCL',
+    'Tesla': 'TSLA',
+    'Twitter': 'TWTR'
+}
+
+CRYPTO = {
+    'Bitcoin': 'BTC',
+    'EOS': 'EOS',
+    'Ethereum': 'ETH',
+    'NEO': 'NEO',
+    'Ripple': 'XRP'
+}
+
+TECHNICAL = {
+    'technical': 'TCHN'
+}
+
+MARKETS = {
+    'United States Dollar': 'USD'
+}
