+++ Heroku_app/styles.py
@@ -1,0 +1,51 @@
+tabs_styles = {
+    'height': '50px',
+    'vertical-align': 'middle'
+}
+
+colors = {
+    'background': '#111111',
+    'text': '#2EACAC',
+    'plot_bgcolor': '#202020'
+}
+
+tab_style = {
+    'borderBottom': '1px solid #d6d6d6',
+    'vertical-align': 'middle',
+    'backgroundColor': colors['background'],
+    'color': colors['text'],
+    'padding': '6px',
+    'font-size': '2rem',
+}
+
+tab_selected_style = {
+    'borderTop': '2px solid #1d6791',
+    'vertical-align': 'middle',
+    'backgroundColor': '#2EACAC',
+    'color': colors['background'],
+    'padding': '6px',
+    'font-size': '2rem',
+}
+
+style_header = {
+    'backgroundColor': '#2EACAC',
+    'color': '#000000'
+                }
+
+style_cell = {
+    'backgroundColor': 'rgb(0, 0, 0)',
+    'color': 'white',
+    'textAlign': 'center'
+             }
+
+style_filter = {
+    'backgroundColor': '#2EACAC',
+}
+
+style_legend = dict(
+    orientation="h",
+    yanchor="top",
+    y=1.1,
+    xanchor="center",
+    x=0.5
+)
